do mobile detection in webdriver package
