+++ packages/webdriver/src/utils.js
@@ -210,3 +210,33 @@ export function isChromiumSession (capabilities) {
         Boolean(capabilities['goog:chromeOptions'])
     )
 }
+
+/**
+ * check if current platform is mobile device
+ *
+ * @param  {Object}  caps  capabilities
+ * @return {Boolean}       true if platform is mobile device
+ */
+export function mobileDetector (caps) {
+    let isMobile = Boolean(
+        (typeof caps['appium-version'] !== 'undefined') ||
+        (typeof caps['device-type'] !== 'undefined') || (typeof caps['deviceType'] !== 'undefined') ||
+        (typeof caps['device-orientation'] !== 'undefined') || (typeof caps['deviceOrientation'] !== 'undefined') ||
+        (typeof caps.deviceName !== 'undefined') ||
+        // Check browserName for specific values
+        (caps.browserName === '' ||
+             (caps.browserName !== undefined && (caps.browserName.toLowerCase() === 'ipad' || caps.browserName.toLowerCase() === 'iphone' || caps.browserName.toLowerCase() === 'android')))
+    )
+
+    let isIOS = Boolean(
+        (caps.platformName && caps.platformName.match(/iOS/i)) ||
+        (caps.deviceName && caps.deviceName.match(/(iPad|iPhone)/i))
+    )
+
+    let isAndroid = Boolean(
+        (caps.platformName && caps.platformName.match(/Android/i)) ||
+        (caps.browserName && caps.browserName.match(/Android/i))
+    )
+
+    return { isMobile, isIOS, isAndroid }
+}
