+++ packages/webdriver/tests/utils.test.js
@@ -1,6 +1,6 @@
 import {
     isSuccessfulResponse, isValidParameter, getArgumentType, getPrototype, commandCallStructure, isW3CSession,
-    isChromiumSession
+    isChromiumSession, mobileDetector
 } from '../src/utils'
 
 import appiumResponse from './__fixtures__/appium.response.json'
@@ -115,5 +115,62 @@ describe('utils', () => {
         expect(isChromiumSession(appiumResponse.value.capabilities)).toBe(false)
         expect(isChromiumSession(chromedriverResponse.value)).toBe(true)
         expect(isChromiumSession(geckodriverResponse.value.capabilities)).toBe(false)
+    })
+
+    describe('mobileDetector', () => {
+        it('should not detect mobile app for browserName===undefined', function () {
+            const {isMobile, isIOS, isAndroid} = mobileDetector({})
+            expect(isMobile).toEqual(false)
+            expect(isIOS).toEqual(false)
+            expect(isAndroid).toEqual(false)
+        })
+
+        it('should not detect mobile app for browserName==="firefox"', function () {
+            const {isMobile, isIOS, isAndroid} = mobileDetector({browserName: 'firefox'})
+            expect(isMobile).toEqual(false)
+            expect(isIOS).toEqual(false)
+            expect(isAndroid).toEqual(false)
+        })
+
+        it('should not detect mobile app for browserName==="chrome"', function () {
+            const {isMobile, isIOS, isAndroid} = mobileDetector({browserName: 'chrome'})
+            expect(isMobile).toEqual(false)
+            expect(isIOS).toEqual(false)
+            expect(isAndroid).toEqual(false)
+        })
+
+        it('should detect mobile app for browserName===""', function () {
+            const {isMobile, isIOS, isAndroid} = mobileDetector({browserName: ''})
+            expect(isMobile).toEqual(true)
+            expect(isIOS).toEqual(false)
+            expect(isAndroid).toEqual(false)
+        })
+
+        it('should detect Android mobile app', function () {
+            const {isMobile, isIOS, isAndroid} = mobileDetector({
+                platformName: 'Android',
+                platformVersion: '4.4',
+                deviceName: 'LGVS450PP2a16334',
+                app: 'foo.apk'
+            })
+            expect(isMobile).toEqual(true)
+            expect(isIOS).toEqual(false)
+            expect(isAndroid).toEqual(true)
+        })
+
+        it('should detect Android mobile app without upload', function () {
+            const {isMobile, isIOS, isAndroid} = mobileDetector({
+                platformName: 'Android',
+                platformVersion: '4.4',
+                deviceName: 'LGVS450PP2a16334',
+                appPackage: 'com.example',
+                appActivity: 'com.example.gui.LauncherActivity',
+                noReset: true,
+                appWaitActivity: 'com.example.gui.LauncherActivity'
+            })
+            expect(isMobile).toEqual(true)
+            expect(isIOS).toEqual(false)
+            expect(isAndroid).toEqual(true)
+        })
     })
 })
